Remove as much Collect specific code form PermissionUtils as possible
We're going to need to share this between feature modules
