+++ collect_app/src/main/java/org/odk/collect/android/injection/config/AppDependencyModule.java
@@ -183,7 +183,7 @@ public class AppDependencyModule {
 
     @Provides
     public PermissionUtils providesPermissionUtils() {
-        return new PermissionUtils();
+        return new PermissionUtils(R.style.Theme_Collect_Dialog_PermissionAlert);
     }
 
     @Provides
