+++ collect_app/src/test/java/org/odk/collect/android/activities/FormDownloadListActivityTest.java
@@ -101,6 +101,10 @@ public class FormDownloadListActivityTest {
     }
 
     private static class AlwaysGrantStoragePermissionsPermissionUtils extends PermissionUtils {
+        private AlwaysGrantStoragePermissionsPermissionUtils() {
+            super(R.style.Theme_Collect_Dialog_PermissionAlert);
+        }
+
         @Override
         public void requestStoragePermissions(Activity activity, @NonNull PermissionListener action) {
             action.granted();
